requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==24.1.0
    # via aiogram
aiogram==3.19.0
    # via -r requirements.in
aiohttp==3.9.3
    # via
    #   -r requirements.in
    #   aiogram
aioschedule==0.5.2
    # via -r requirements.in
aiosignal==1.3.1
    # via
    #   -r requirements.in
    #   aiohttp
aiosqlite==0.21.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
apscheduler==3.11.0
    # via -r requirements.in
async-timeout==4.0.3
    # via aiohttp
attrs==25.3.0
    # via aiohttp
certifi==2025.1.31
    # via
    #   aiogram
    #   requests
charset-normalizer==3.4.1
    # via requests
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via
    #   requests
    #   yarl
magic-filter==1.0.12
    # via aiogram
multidict==6.2.0
    # via
    #   aiohttp
    #   yarl
propcache==0.3.1
    # via yarl
pydantic==2.10.6
    # via aiogram
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
typing-extensions==4.13.0
    # via
    #   aiogram
    #   aiosqlite
    #   multidict
    #   pydantic
    #   pydantic-core
tzdata==2025.2
    # via tzlocal
tzlocal==5.3.1
    # via apscheduler
urllib3==2.3.0
    # via requests
yarl==1.18.3
    # via aiohttp
